Add copyToClipboard to helpers file
