--- assets/src/js/utils/helpers.js
@@ -1,8 +1,0 @@
-export const copyToClipboard = ( text ) => {
-	const temp = document.createElement( 'input' );
-	document.body.appendChild( temp );
-	temp.value = text;
-	temp.select();
-	document.execCommand( 'copy' );
-	temp.remove();
-};
